move to agorapulse and add kotest
