+++ starter-core/src/main/java/io/micronaut/starter/feature/agorapulse/gru/template/helloWorldGruControllerGroovy.rocker.raw
