+++ starter-core/src/main/java/io/micronaut/starter/feature/agorapulse/gru/template/helloWorldGruControllerKotlin.rocker.raw
