+++ starter-core/src/main/java/io/micronaut/starter/feature/agorapulse/gru/template/helloWorldGruControllerTestKotest.rocker.raw
@@ -1,0 +1,24 @@
+@import io.micronaut.starter.application.Project
+
+@args (
+Project project
+)
+
+@if (project.getPackageName() != null) {
+package @project.getPackageName()
+}
+
+import com.agorapulse.gru.Gru
+import io.micronaut.test.extensions.kotest.annotation.MicronautTest
+import io.kotest.core.spec.style.StringSpec
+
+@@MicronautTest
+class HelloWorldGruControllerTest(private val gru: Gru): StringSpec({
+
+    "get hello message" {
+        gru.verify { it
+            .get("/gru")
+            .expect { it.json("gruIndex.json") }
+        }
+    }
+})
