+++ starter-core/src/main/java/io/micronaut/starter/feature/agorapulse/gru/template/helloWorldGruControllerTestKotlin.rocker.raw
